Fix configure on GNU Hurd
PATH_MAX is not required by POSIX, and it's not defined on GNU Hurd:
https://www.gnu.org/software/hurd/community/gsoc/project_ideas/maxpath.html

In this case the fix is easy, as PATH_MAX is only used for reserving some space in a string, which is just a performance optimization. So it can just be skipped in case PATH_MAX is not defined.